toxiverse/requirements.txt
Flask==2.2
gunicorn==20.1.0
pandas==1.5.2
pip==22.3.1
plotly==5.9.0
rdkit==2022.09.2
flask-sqlalchemy==3.0.2
flask-login==0.6.2
flask-migrate==4.0.0
Flask-Mail==0.9.1
scikit-learn==1.2.0
Werkzeug==2.2.2
redis==4.4.0
rq==1.11.1
PyJWT==2.6.0
numpy==1.23.2
requests==2.30.0
python-dotenv==1.0.0
matplotlib==3.7.0
seaborn==0.13.0
wheel==0.41.3
alembic==1.12.1
tensorflow==2.10.0
typing-extensions==4.2.0
APScheduler>=3.10.0
